requirements.txt
Flask==2.0.1
flask_jwt_extended==4.5.2
Flask-WTF==0.15.1
WTForms==3.0.0
python-dotenv==0.19.1
passlib==1.7.4
bootstrap-flask==2.2.0